using tabs, fixing crashes, fixing bugs
Fixed a crash caused by all of these conditions being true
* global variable is declared before crash location
* global variable gets something added to it in a sepeate scope, like an other function
* added thing gest accessed or modified.
* cursor is after the access or modification

thanks to v6 for sending me a file in which i coule relyably test. and fix that error.

switched the package from using 2 space = 1 indent to 1tab = 1 indent, because tabs are meant for that.

Fixed a bug/scoping annoyance, where you would not see the names / specifications of the functions and variable being created after the current cursor position.
Silverfeelin notified me about it and well it's now fixed, so enjoy

changed the default option to not show the default auto-complete as most of them are either redundent or clutter.

fixed a bug where the includeSuggestion setting would not affect anything, this was due to some old residual from the time I turned this package into it's own thing.