dissabling rc caching, expanded documentation formats, fixed some formatting
Fixed the spacing in the formatting, god knows how i missed that one....

Dissabled the rc caching as it seems to fix some bugs with custom rcs getting overriden.

Added a way to display return types for varients in main and rc documentation

Started porting more of the vannila lua to the dynamic system used to display types and create snippets.